Improve auto locking
Changed to using xidlehook instead of xautolock.

xidlehook --not-when-audio depends on pulseaudio, so a custom bash
script .yadm/bin/suspend-if-no-audio has been added that works with
alsa.

xidlehook --not-when-fullscreen does not seem to be working for
unknown reason. It has been included in the command anyway in
case a future update fixes it.